move youtube_search.py, add class for search results
